Task-Induced Activation Intensity/requirements.txt
torch>=2.0.0
transformers>=4.40.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.3.0
scipy>=1.11.0
safetensors>=0.4.0
accelerate>=0.20.0

